requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.16.0
webdriver-manager==4.0.1
flask==3.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pandas==2.1.3
lxml==5.0.0
gunicorn==21.2.0
